Excluded lexical and its dependencies from completion
Since lexical partially pollutes the project's namespace with the
dependencies that RemoteControl brings in, it's prudent to exclude
those, and only those dependencies from completion.
However, if you're working on one of those projects, you probably want
to see those dependencies, so lexical checks to see if that's the
case, and lets them in if so.