Moved project test fixtures to remote control (#39)
* Moved project test fixtures to remote control

* Lexical coludn't format some of lexical's projects

I'm pretty sure these fixes are specific to lexical, but getting the
formatter.exs was failing in remote control because elixir said that
proto wasn't defined, but that's odd, since the .formatter.exs in
remote control didn't import that dep. Setting the dep to optional
fixed it.

I also found a case where formatting mix.exs would fail because a
diagnostic context wasn't found, so I defaulted the position to 0

* Error was not the correct value

should have been :error, not error.

* Fixed error about mix project being redefined

If we detect that we're compiling a mix.exs file, we can pop the
current project off the stack and then push it back on, which makes
the error go away.