Module sorting / Refactor boost (#212)
* Module sorting / Refactor boost

This pr does two things:

 1. Reorders completions so that modules come before functions
 2. Refactors boost so that it operates on Completion.Item structs

I think the first thing is a welcome change; it's the default for most
LS implementations to sort Modules before functions, and the makes
boosting things easier. All items now have a default boost of 0, and
items can have a local boost with a number from 0-9 to boost items of the same
kind, or a global boost to boost across item kinds.