Reworked lexical packaging
This commit moves away from releases for our packaging. What replaces
them is a package directory that has a launcher, working config and
priv directories and a bunch of erlang archives for storing our beam
files.

New packaging enables the following:

  * We can compile on one version of erlang and run on another
  * We get a real priv directory experience, which isn't possible with
    escripts.
  * We get a "real" config directory
  * We get a boot script that's just a plain old elixir file
  * All of our files can be ported to windows fairly easily
  * No tricks for packaging scripts / config. They're just plain
    files, and we rely on the `:code` module to load them

I've tested this compiling in one version of elixir and erlang and
running it in another, and it seems to work great. It makes sense
though, they're just BEAM files, which should be cross platform.

Other approaches:

We all know why releases won't / don't work. I also tried building an
escript, and it kind of worked, but had the following drawbacks:

  * There is no priv directory. I was able to code around this by
    having a module that generates the port_mapper.sh file for you,
    but this was a big cumbersome.
  * An escript is a binary file, and erlang loads all of its modules
    from it. This was a problem for the second VM, as we couldn't just
    point it at the binary file and have it look for modules there. I
    also tried using the 'inet' loader, but that had a bunch of
    extremely difficult to debug issues. Worse, I would still need
    some erlang archives, so why not just make it universal?

I also tried using `mix archive.build`, but it didn't support
namespacing (obviously) and changing the task to do so would pretty
much lead us to the path we have here.

Fixes #255