0.3.0 Release (#337)
* Bumped versions to 0.3.0

* Added version compatibility check (#332)

* Added version compatibility check

Lexical now outputs the erlang and elixir versions of the VM that was
used to create them. It now checks the build and packaging directories
for these versioning files and emits an error message if the vm starts
with beam files that won't work.

* Detected version change on call to package

Another belt to wear with the suspenders. When we build a package that
overwrites another package, if the version has changed, we delete the
old compiled code and the old package and start compilation from
scratch. This definitely prevents old beam files from ending up in the
new package.

* Fixed dialyzer error

halt/1 always raises which irritates dialyzer. Added an ignore to it.

* Improved launcher script (#339)

* Improved launcher script

Seems ERL_LIBS wasn't picked up universally across versions, so I
swapped that out for the -pa flag

* silenced errors, used command -v to check for asdf / rtx

* Evalled output from rtx env

* fixed typo

* Update bin/start_lexical.sh

Co-authored-by: Étienne Lévesque <contact@etiennel.dev>

---------

Co-authored-by: Étienne Lévesque <contact@etiennel.dev>

* fix: update Nix derivation to use new build (#344)

* fix: update Nix derivation to use new build

---------

Co-authored-by: Steve Cohen <scohen@scohen.org>

---------

Co-authored-by: Étienne Lévesque <contact@etiennel.dev>
Co-authored-by: Łukasz Jan Niemier <lukasz@niemier.pl>