Centralize general-use AST modules in `common` (#342)
* Minor cleanup in completion env

* Extract completions env into `Lexical.Ast.Env`

Completions are currently built on top of two behaviours, `Environment` and
`Builder`. The first describes a behaviour for inspecting the contextual
environment of a given position so that you can filter completions to only
those that would make sense in that context. The second describes a behaviour
for turning those completions into something that the LSP understands.

The `Environment` behaviour and implementation is generally useful outside of
the context of completions. The completions feature implemented both behaviours
in a single `Lexical.Server.CodeIntelligence.Completion.Env`. This commit
decouples the env and the builder implementations, moving the
environment-related code (and struct) to `Lexical.Ast.Env` and putting the
builder implementation in `Lexical.Server.CodeIntelligence.Completion.Builder`.

* Move `RemoteControl.CodeMod.Ast.*` into `Ast`

`Lexical.RemoteControl.CodeMod.Ast` contained generally-useful functions for
parsing and creating zippers from documents. While it was located in
`RemoteControl`, it wasn't doing anything that required running on the remote
node.

Similarly, `Lexical.RemoteControl.CodeMod.Ast.Aliases` contains code for
resolving the available aliases at a given position and also has no
requirements to run on the remote node. It's been moved to
`Lexical.Ast.Aliases`.

* Move `RemoteControl.CodeIntelligence.Ast` into `Ast`

This module provided only a relatively thin wrapper over `Aliases.at/2`. It has
been moved into `Lexical.Ast`.

* Delete outdated comment

This was a comment explaining why position line and character were being
incremented when they used to be zero-based. They are now one-based, so
this comment is no longer needed.

* Add docs and typespecs to `Lexical.Ast`