Fixing crash when dealing with unicode files (#672)
The change to binread introduced a fairly complicated unicode bug that
would cause decoding of messages to partially fail. This was due to
the fact that we were previously using read, which would occasionally
return unencoded data. We patched this by doing a utf8 -> latin1
conversion, but this would fail once we acutally read the bytes rather
than going through elixir's utf8 encoding friendly IO.read functions.

The change here is to get rid of the encoding step, as we're reading
the raw bytes as we should.