Included non-deps symbols in fuzzy matcher (#704)
* Included non-deps symbols in fuzzy matcher

I noticed that the fuzzy matcher wouldn't match any unit tests. This
is because their application is nil, as they're scripts.
This change checks for nil applications, and if it finds one, only
excludes the mapped entry if the path looks like it comes from a
dependencies directory.

Note, the deps directory name is hard coded. I originally queried the
project for the deps directory name, but this would include
dependencies in the projects directory, which was quite annoying,
hence the hard-coding.

* Changed how we detect dependencies

Dependencies are now detected via seeing if the path is in our deps
directory or one of our sub-project's dependencies.


Co-authored-by: Zach Allaun <zach.allaun@gmail.com>

---------

Co-authored-by: Zach Allaun <zach.allaun@gmail.com>