Update test that fails after the rename
There is an existing bug in `main` that causes `ExUnitProperties` to be
suggested for `Exper|t`.

The changes in the PR this commit is part of is focused only in the renaming,
so this changes the test to something that tests roughly the same, but passes.

A different PR needs to fix the underlying issue here